+++ test/fifo_test.c
@@ -34,7 +34,7 @@ static bool has_sigpipe(void) {
 static bool reader_open_finished = false;
 static void* do_reader_open(void* arg) {
   apos_fifo_t* fifo = (apos_fifo_t*)arg;
-  int result = fifo_open(fifo, FIFO_READ, true, false);
+  intptr_t result = fifo_open(fifo, FIFO_READ, true, false);
   reader_open_finished = true;
   return (void*)result;
 }
@@ -42,7 +42,7 @@ static void* do_reader_open(void* arg) {
 static bool writer_open_finished = false;
 static void* do_writer_open(void* arg) {
   apos_fifo_t* fifo = (apos_fifo_t*)arg;
-  int result = fifo_open(fifo, FIFO_WRITE, true, false);
+  intptr_t result = fifo_open(fifo, FIFO_WRITE, true, false);
   writer_open_finished = true;
   return (void*)result;
 }
@@ -138,7 +138,7 @@ static void open_test(void) {
 
   KEXPECT_EQ(0, fifo_open(&fifo, FIFO_WRITE, true, false));
   KEXPECT_EQ(1, kthread_queue_empty(&fifo.read_queue));
-  KEXPECT_EQ(0, (int)kthread_join(thread));
+  KEXPECT_EQ(0, (intptr_t)kthread_join(thread));
   KEXPECT_EQ(true, reader_open_finished);
   fifo_close(&fifo, FIFO_READ);
   fifo_close(&fifo, FIFO_READ);
@@ -168,7 +168,7 @@ static void open_test(void) {
 
   KEXPECT_EQ(0, fifo_open(&fifo, FIFO_READ, true, false));
   KEXPECT_EQ(1, kthread_queue_empty(&fifo.write_queue));
-  KEXPECT_EQ(0, (int)kthread_join(thread));
+  KEXPECT_EQ(0, (intptr_t)kthread_join(thread));
   KEXPECT_EQ(true, writer_open_finished);
   fifo_close(&fifo, FIFO_READ);
   fifo_close(&fifo, FIFO_WRITE);
